src/requirements.txt
fastapi[standard]~=0.115.12
pytest~=8.3.5
requests~=2.32.3
uvicorn~=0.34.2
pydantic~=2.11.4
starlette~=0.46.2
pydantic-settings~=2.9.1
pandas~=2.3.0
cryptography
pyjwt[crypto]
xmltodict